refactor: move DiffraxIntegrator
